requirements.txt
asgiref==3.11.0
boto3==1.42.23
botocore==1.42.23
Django==6.0
django-appconf==1.2.0
django-imagekit==6.0.0
django-storages==1.14.6
django_urls==1.1.4
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pilkit==3.0
pillow==12.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.2
